Added back FixedDecimal and UInt256 validations where necessary

+++ test/Opdex.Platform.Application.Tests/EntryHandlers/Transactions/TransactionLogs/Markets/ProcessChangeMarketPermissionLogCommandHandlerTests.cs
@@ -131,7 +131,7 @@ namespace Opdex.Platform.Application.Tests.EntryHandlers.Transactions.Transactio
         {
             dynamic data = new ExpandoObject();
             data.address = address.ToString();
-            data.permission = permission;
+            data.permission = (byte)permission;
             data.isAuthorized = isAuthorized;
             return data;
         }
